+++ include/entities/elemental-attributes.h
@@ -16,8 +16,9 @@ public:
 
 // private:
     static std::vector<std::pair<std::string, int> > elementalAttributes;
-    //single elemental attributes
-    static std::vector<std::vector<int> > elementalAttributesRestraints;
+    // 克制倍率矩阵：存官方原始倍率（0=免疫 / 0.5=减半 / 1.0=普通 / 2.0=克制）。
+    // 用 double 而非 int 三档编码——0.5 减半语义必须保留（旧编码把 0.5 丢成 0=免疫）。
+    static std::vector<std::vector<double> > elementalAttributesRestraints;
 };
 
 #endif // ELEMENTAL_ATTRIBUTES_H
